In "./configure" of HPLIP tell the build architecture explicitly
HPLIP seems to use a stone-old config-guess script which discovers ARM
architectures as "aarch64" and claims to be from 2011, which makes the
Snap build fail on these architectures. Debian packaging generally
uses the "--build=..." argument when calling "./configure" to manually
tell the build architecture.

Now we do this for the Snap build of HPLIP, too.