Fix Snap builds on non-amd64 architectures
HPLIP provides an optionally usable binary-only blob for its "hpcups"
printer driver filter, libImageProcessor, which is only available for
i386 (not supported by core20) and amd64.

It does not automatically enables or disables it depending on the
build architecture, but only manually enabling or disabling it by a
command line argument for "./configure".

To allow building the Snap on all supported architectures without
generally disabling this feature we auto-select in the Snap build, via
"build-environment" with advanced grammar now.