Snap build: Make "ppdc" in HPLIP's "./configure" work
The "./configure" process of HPLIP runs CUPS' "ppdc" on a the
prnt/drv/hpcups.drv file of HPLIP's source code, to update the
included PPD files.

We have built the "ppdc" in the "cups" part, but snapcraft only
supports linkning to shared libraries of another part by default, not
running programs of another part. So we had to manually set LD_LIBRARY
path in the "hplip" part to the /usr/lib directory in the stage
directory.

In addition we have to stage CUPS' /usr/share/cups/ppdc/*.defs files
and set the CUPS_DATADIR environment variable in the "hplip" part.

The updating of the PPD files by "ppdc" is especially important if we
apply patches to the source code of HPLIP before building it.