Added PostScript printer support
Now this Printer Application can be used for printing on all printers
supported by HPLIP (without proprietary plugin for now), both
non-PostScript and PostScript printers.

The Snap now contains also the PostScript PPD files and the `hpps`
filter for PIN-protected printing on PostScript printers. The
Ghostscript in the Snap has the "ps2write" PostScript output device
included for converting PDF input into PostScript.

The Printer Application's main executable has conversion rules for
PostScript added and also allows considering and listing PostScript
PPD entries on the "Add Printer" web interface page and in the output
of "drivers" sub-command.

Man page and README.md are also appropriately updated.

In addition, to improve convenience for developers (and everyone who
wants to build from source), we do a "make clean" before "make" in the
"hplip-printer-app" part in snapcraft.yaml, because if we had done
"make" off-Snap, directly in the source tree. and afterwards build the
Snap with snapcraft, the build sucks in our local binary of
hplip-printer-app instead of compiling its own one in the Snap harness
with the appropriate libraries, ending up with the Snap containing an
executable which does not work inside the Snap. The "make clean"
removes any accidentally grabbed binary.