Fixed Copy&Paste error in path for "pdftops"
Usually, the Printer Applications use Ghostscript, called by the
pdftops() filter function, to convert PDF input into PostScript, but
there are some printers with PostScript interpreter bugs which do not
print the PostScript produced by Ghostscript. For these pdftops()
applies some quirk rules to either fix Ghostscript's PostScript, apply
workarounds to it, or it swirtches to Poppler, using Poppler's
"pdftops" utility.

In the Snap this utility is not in the usual /usr/bin directory,
therefore we supply its actual place in the "./configure" call for the
building of cups-filters. Here we had a Copy&Paste error making the
utility not being found and discovered it only recently by a user with
one of the quirky printers, the HP LaserJet 4050 (PostScript mode).

See attached log of

https://github.com/OpenPrinting/ps-printer-app/issues/9#issuecomment-915186577