Switched from HP's Upstream source to Debian's package source
As HPLIP has many bugs which got reported by distro maintainers but
not taken care of by HP's HPLIP developers, fixes got applied\as
patches in the ditro packages, 81 (!)  in the Debian package (which is
also used by Ubuntu).

Creating the Snap with the original source code from HP and not
applying all these patches would leave the HPLIP Printer Application
as buggy as the original HPLIP, but separately maintaining these
patches for the Snap, and especially if one tries to find out which
patches are the relevant ones for the Printer Application and which
not, would be a lot of unneeded work. Also having to updated the
patches when the Snap is updated to a new upstream source version of
HPLIP.

So now, with Debian's new printing maintainer, Thorsten Alteholz,
having picked up and updated Debian's HPLIP package to 3.21.8 (thanks
a lot!) we switch over to using the source of HPLIP from Debian for
the Snap, with all patches applied.

The code is downloaded from Debian's GitLab site:

https://salsa.debian.org/printing-team/hplip/

Currently, to grab 3.21.8+dfsg0-2 the tag debian/3.21.8+dfsg0-2 is
downloaded.

Some changes in the build had to be done:

- The binary-only ImageProcessor library is not contained in Debian's
  source code and so we do not need to handle it.
- In Debian's source the included PPD files and some data files are
  uncompressed fo better GIT tracking, HP's original source has them
  gzipped. So we need to re-compress before building.
- We need to generate the *.drv files after "./configure" and before
  "make" using the Python script "dat2drv.py" to prevent "make" from
  trying to build and run the C++ counterpart "Dat2drv", which does
  not build due to missing Common.h. This requires us to run
  "./configure", "make", and "make install" explicitly, instead of
  running "snapcraftctl build".
- Before "./configure" we run "debian/autogen.sh" to apply the changes
  to the build system in the patches.
- "Make install" installs the Python extension modules ("cupsext.so",
  ...) to the wrong place, where they do not get found inside the
  Snap. This we have to correct now.
- Install the PostScript PPD files, which "make install" misses.