Add support for the proprietary plug-in to the Snap
This commit is mainly to save work but also a proof of concept that
HP's proprietary plugin for HPLIP also work if HPLIP is running in a
fully confined Snap. At least I could get automatic firmware upload
and printing with printer driver extensions to work. Scanner driver
extensions and additional SANE backends in the plugin will probably
also work, but I could not test, as the SANE-driver retro-fitting
framework (aka scanning support in retro-fitting Printer Applications)
is not ready yet.

The commit only contains the parts for using an installed plug-in,
meaning the dynamic link libraries to extend the drivers get used and
firmware files automatically uploaded to printers which need them. It
does not contain the user interface (web interface page) to download
the plugin from HP's official locations and install it into the
Printer Application.

One can test already by manually installing the plugin as follows:

Go to

    https://developers.hp.com/hp-linux-imaging-and-printing/plugins

and download the plugin for the Snap's HPLIP version (currently
3.21.8). Uncompress it via

    sh hplip-3.21.8-plugin.run --noexec --keep

and copy the files into place

    sudo rm -rf /var/snap/hplip-printer-app/common/plugin
    sudo cp -r plugin_tmp /var/snap/hplip-printer-app/common/plugin

In the plugin directory you have to link the *.so of your architecture
to the names without architecture specification:

On amd64 (Intel/AMD 64-bit, x86_64):

    cd /var/snap/hplip-printer-app/common/plugin
    sudo ln -s lj-x86_64.so lj.so
    sudo ln -s hbpl1-x86_64.so hbpl1.so

On armhf (ARM 32-bit):

    cd /var/snap/hplip-printer-app/common/plugin
    sudo ln -s lj-arm32.so lj.so
    sudo ln -s hbpl1-arm32.so hbpl1.so

On arm64 (ARM 64-bit):

    cd /var/snap/hplip-printer-app/common/plugin
    sudo ln -s lj-arm64.so lj.so
    sudo ln -s hbpl1-arm64.so hbpl1.so

Theoretically you have to do it with all *.so files, but as we only
support printing (and not scanning) currently, this is enough for your
testing.

Now create the plugin state file

    /var/snap/hplip-printer-app/common/var/hplip.state

to contain

    [plugin]
    installed = 1
    eula = 1
    version = 3.21.8

Restart the Snap

    sudo snap stop hplip-printer-app
    sudo snap start hplip-printer-app

Now you can print also on the printers marked with "requires
proprietary plugin" in the driver list of the "Add Printer" web
interface page and in the entries of the printers set up in the web
interface.

If the printer needs its firmware loaded after being turned on, this
is done automatically with the plugin once installed.

Note that the plugin has a different license than HPLIP itself:

    /var/snap/hplip-printer-app/common/plugin/license.txt

The next step of the development of this Printer Application will be
the addition of a web interface to install the plugin. This will also
contain checksum/signature verification, presentation of the license,
and automatic update when the HPLIP version gets updated.

The use of the printer driver extensions of the plugin is simply done
by adjusting the paths to find the plugin's dynamic link libraries in
HPLIP's printer driver.

For the automatic firmware uploading the classic HPLIP installation
installs UDEV rules into the system so that if a USB printer from HP
gets connected the HPLIP utility to upload the firmware gets
started. The utility then checks whether the printer actually needs
its firmware loaded and uploads it if it is needed and the firmware
file (from the plugin) is present.

A Snap is not capable of installing UDEV rules into the system, so it
uses a different method. If the plugin is installed, it runs an
auxiliary daemon running "udevadm monitor" to observe the
(dis)appearing of USB devices and checks whether an event is from an
appearing HP printer and then triggers the firmware utility. The
daemon is actually a forking part in the

    snap/local/run-hplip-printer-app-server

schell script.

The same method is also used in the ipp-usb Snap:

    https://github.com/OpenPrinting/ipp-usb

in the

    snap/local/run-ipp-usb

file, there to start ipp-usb on appearing USB printers with
IPP-over-USB capability.