Add support for the proprietary plug-in to the Snap - Installation
In the first commit we made the HPLIP Printer Application, both
classically installed and in a fully confined Snap, supporting the
proprietary plugin from HP, but to use it, we still had to install it
manually.

This commit adds automatic installation functionality for the plugin
to the Printer Application. Once the plugin manually installed, it
will get automatically updated when HPLIP is replaced by a newer
upstream version. Especially an automatic update of the HPLIP Printer
Application Snap by a one with a newer version of HPLIP would also
update the plugin, to make sure that it always works.

The installation procedure of the Printer Application uses the same
security features as HP's installation scripts in HPLIP, checking the
size, SHA1 checksum, and GPG signature, so an installation of a
malicious fake plugin should not happen.

Important to say is that the automatic installation/update of the
plugin only works when the Printer Application is running as root (in
the Snap it is doing so).

Note that the plugin support is not yet complete, as the web interface
page for installing the plugin easily is not yet added. We have merely
another proof of concept, this time showing that we can install the
plugin out of a fully confined Snap.

Main addition are the functions to do the plugin installation in
hplip-printer-app.c. They are made in a way that they work both for a
classic installation of the Printer Application together with a
classically installed HPLIP (it auto-updates the plugin also in this
configuration) and for the fully confined Snap. Conditional
compilation is used to fit the Printer Application to the different
configurations.

In addition, the startup script for the server is improved to check
for the presence of the plugin every 5 seconds and as soon as it
appears, starts USB monitoring for firmware uploads. Before, there was
only one check in the very beginning, meaning that one had to restart
the Printer Application after installing the plugin (and also after an
automatic update of the plugin) to get the plugin working.

For testing one can trigger the installation process by removing the
plugin state file (if there is already one), putting an empty file in
its place, and restart the Printer Application. As "cheat mode" for
testing until we have the web interface page in place an empty plugin
state file also triggers an automatic installation of the plugin, not
only an outdated/non-matching version number in the state file.

So

    sudo rm -f /var/snap/hplip-printer-app/common/var/hplip.state
    sudo touch /var/snap/hplip-printer-app/common/var/hplip.state
    sudo snap stop hplip-printer-app
    sudo snap start hplip-printer-app

triggers the installation of the plugin.

The same also works for the classic installation (/var/lib/hp/hplip.state).