Add support for the proprietary plug-in to the Snap - Web Interface
In the first commit we made the HPLIP Printer Application, both
classically installed and in a fully confined Snap, supporting the
proprietary plugin from HP, and in the second commit we added the
functionality for installing and auto-updating the plugin, but no user
interface to control the installation or removal.

This commit now completes the support for HP's proprietary plugin by
adding a user interface for managing it. Now there is a button
"Install Proprietary Plugin" on the front page of the web interface,
in the "Other Settings" section on the left, and also the "Plugin"
buttons in the entries of each printer which needs the plugin.

These buttons open the web interface to manage the installation of the
plugin. It explains for what the plugin is good for, shows its
installation status: Not installed, installed and up-to-date, or
out-of-date, two buttons, one to install/re-install/update and one to
remove the plug-in, and finally the license of the plugin (only if it
is installed).

Removing the plugin is only available in the Snap of this Printer
Application, not in the classic installation, as the installation of
the plugin into a classically installed HPLIP is rather complex and HP
does not provide an uninstall utility, whereas uninstalling in the
Snap is simply removing the plugin directory.

On re-install and remove there is shown an "Are you sure?" page before
the actual action is taken and on initial installation, the license
text is shown and only if the user agrees, the actual installation
takes place and on decliningthe license, the downloaded plugin is
removed again.

We also check now whether the Printer Application is running as root
and only then we do any installation, update, or removal action of the
plugin and also only then we show appropriate buttons in the web
interface. In the Snap the Printer Application is always running as
root, in the classic installation, with the Printer Application not
running as root, the user is supposed to run the "hp-plugin" utility
of the classically installed HPLIP (the Printer Application shows the
status of the plugin though).

In addition, some typos were fixed.