Set longer systemd timeout for the Printer Application
Use a systemd timeout longer than 1 min as PAPPL has a 1-minute
shutdown timeout if there are pending jobs for which the printer
is not responding. This way we avoid the Printer Application
getting killed and shut down uncleanly.

Using "stop-timeout: 70s" in the apps: entry for the server in
snapcraft.yaml.

Solves

    https://github.com/OpenPrinting/ghostscript-printer-app/issues/4

for all the PAPPL-based Printer Applications.